global variables are moved to separate file
