+++ client/block.c
@@ -10,6 +10,7 @@
 #include "../ldus/source/include/ldus/rbtree.h"
 #include "block.h"
 #include "crypt.h"
+#include "global.h"
 #include "wallet.h"
 #include "storage.h"
 #include "transport.h"
@@ -114,7 +115,6 @@ static struct cache_block *cache_first = NULL, *cache_last = NULL;
 static pthread_mutex_t block_mutex;
 static pthread_mutex_t rbtree_mutex;
 //TODO: this variable duplicates existing global variable g_is_pool. Probably should be removed
-static int g_light_mode = 0;
 static uint32_t cache_bounded_counter = 0;
 static struct orphan_block *g_orphan_first[ORPHAN_HASH_SIZE], *g_orphan_last[ORPHAN_HASH_SIZE];
 
@@ -542,7 +542,7 @@ static int add_block_nolock(struct xdag_block *newBlock, xtime_t limit)
 		}
 	}
 
-	if(g_light_mode) {
+	if(g_xdag_type == XDAG_WALLET) {
 		outmask = 0;
 	}
 
@@ -561,7 +561,7 @@ static int add_block_nolock(struct xdag_block *newBlock, xtime_t limit)
 	}
 
 	/* if not read from storage and timestamp is ...ffff and last field is nonce then the block is extra */
-	if (!g_light_mode && (transportHeader & (sizeof(struct xdag_block) - 1))
+	if (g_xdag_type == XDAG_POOL && (transportHeader & (sizeof(struct xdag_block) - 1))
 			&& (tmpNodeBlock.time & (MAIN_CHAIN_PERIOD - 1)) == (MAIN_CHAIN_PERIOD - 1)
 			&& (signinmask & 1 << (XDAG_BLOCK_FIELDS - 1))) {
 		tmpNodeBlock.flags |= BI_EXTRA;
@@ -585,7 +585,7 @@ static int add_block_nolock(struct xdag_block *newBlock, xtime_t limit)
 		}
 	}
 
-	if(!g_light_mode) {
+	if(g_xdag_type == XDAG_POOL) {
 		check_new_main();
 	}
 
@@ -839,7 +839,7 @@ void *add_block_callback(void *block, void *data)
 
 	pthread_mutex_unlock(&block_mutex);
 
-	if(res >= 0 && g_xdag_pool) {
+	if(res >= 0 && g_xdag_type == XDAG_POOL) {
 		xdag_sync_pop_block(b);
 	}
 
@@ -928,7 +928,7 @@ struct xdag_block* xdag_create_block(struct xdag_field *fields, int inputsCount,
 	block[0].field[0].time = send_time;
 	block[0].field[0].amount = fee;
 
-	if (g_light_mode) {
+	if (g_xdag_type == XDAG_WALLET) {
 		pthread_mutex_lock(&g_create_block_mutex);
 		if (res < XDAG_BLOCK_FIELDS && ourfirst) {
 			setfld(XDAG_FIELD_OUT, ourfirst->hash, xdag_hashlow_t);
@@ -1142,7 +1142,7 @@ begin:
 
 	// launching of synchronization thread
 	g_xdag_sync_on = 1;
-	if (!g_light_mode && !sync_thread_running) {
+	if (g_xdag_type == XDAG_POOL && !sync_thread_running) {
 		xdag_mess("Starting sync thread...");
 		int err = pthread_create(&th, 0, sync_thread, 0);
 		if(err != 0) {
@@ -1159,7 +1159,7 @@ begin:
 		}
 	}
 
-	if (g_light_mode) {
+	if (g_xdag_type == XDAG_WALLET) {
 		// start mining threads
 		xdag_mess("Starting mining threads...");
 		xdag_mining_start(n_mining_threads);
@@ -1181,7 +1181,7 @@ begin:
 			g_xdag_extstats.hashrate_s = ((double)(nhashes - nhashes0) * 1024) / (t - t0);
 		}
 
-		if (!g_block_production_on && !g_light_mode &&
+		if (!g_block_production_on && g_xdag_type == XDAG_POOL &&
 				(g_xdag_state == XDAG_STATE_WAIT || g_xdag_state == XDAG_STATE_WTST ||
 				g_xdag_state == XDAG_STATE_SYNC || g_xdag_state == XDAG_STATE_STST || 
 				g_xdag_state == XDAG_STATE_CONN || g_xdag_state == XDAG_STATE_CTST)) {
@@ -1240,7 +1240,7 @@ begin:
 			time_t last_received = atomic_load_explicit_uint_least64(&g_xdag_last_received, memory_order_relaxed);
 
 			if (t > (last_received << 10) && t - (last_received << 10) > 3 * MAIN_CHAIN_PERIOD) {
-				g_xdag_state = (g_light_mode ? (g_xdag_testnet ? XDAG_STATE_TTST : XDAG_STATE_TRYP)
+				g_xdag_state = (g_xdag_type == XDAG_WALLET ? (g_xdag_testnet ? XDAG_STATE_TTST : XDAG_STATE_TRYP)
 					: (g_xdag_testnet ? XDAG_STATE_WTST : XDAG_STATE_WAIT));
 				conn_time = sync_time = 0;
 			} else {
@@ -1248,14 +1248,14 @@ begin:
 					conn_time = t;
 				}
 
-				if (!g_light_mode && t - conn_time >= 2 * MAIN_CHAIN_PERIOD
+				if (g_xdag_type == XDAG_POOL && t - conn_time >= 2 * MAIN_CHAIN_PERIOD
 					&& !memcmp(&g_xdag_stats.difficulty, &g_xdag_stats.max_difficulty, sizeof(xdag_diff_t))) {
 					sync_time = t;
 				}
 
 				if (t - (g_xdag_xfer_last << 10) <= 2 * MAIN_CHAIN_PERIOD + 4) {
 					g_xdag_state = XDAG_STATE_XFER;
-				} else if (g_light_mode) {
+				} else if (g_xdag_type == XDAG_WALLET) {
 					g_xdag_state = (g_xdag_mining_threads > 0 ?
 						(g_xdag_testnet ? XDAG_STATE_MTST : XDAG_STATE_MINE)
 						: (g_xdag_testnet ? XDAG_STATE_PTST : XDAG_STATE_POOL));
@@ -1267,7 +1267,7 @@ begin:
 			}
 		}
 
-		if (!g_light_mode) {
+		if (g_xdag_type == XDAG_POOL) {
 			check_new_main();
 		}
 
@@ -1288,16 +1288,12 @@ begin:
  *   for the light node is_pool == 0;
  * miner_address = 1 - the address of the miner is explicitly set
  */
-int xdag_blocks_start(int is_pool, int mining_threads_count, int miner_address)
+int xdag_blocks_start(int mining_threads_count, int miner_address)
 {
 	pthread_mutexattr_t attr;
 	pthread_t th;
 
-	if (!is_pool) {
-		g_light_mode = 1;
-	}
-
-	if (xdag_mem_init(g_light_mode && !miner_address ? 0 : (((xdag_get_xtimestamp() - XDAG_ERA) >> 10) + (uint64_t)365 * 24 * 60 * 60) * 2 * sizeof(struct block_internal))) {
+	if (xdag_mem_init(g_xdag_type == XDAG_WALLET && !miner_address ? 0 : (((xdag_get_xtimestamp() - XDAG_ERA) >> 10) + (uint64_t)365 * 24 * 60 * 60) * 2 * sizeof(struct block_internal))) {
 		return -1;
 	}
