+++ client/storage.c
@@ -6,7 +6,7 @@
 #include <pthread.h>
 #include <unistd.h>
 #include "storage.h"
-#include "init.h"
+#include "global.h"
 #include "hash.h"
 #include "utils/log.h"
 #include "utils/utils.h"
